Move express-server CLI command from terra-open-source-scripts to terra-functional-testing to eliminate the need to depend on terra-open-source-scripts to use express-server.
