+++ packages/terra-open-source-scripts/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Removed
+  * Removed express-server CLI command. This is moved to @cerner/terra-functional-testing to eliminate the need to depend on this package to use express-server.
+
 ## 1.8.0 - (April 23, 2021)
 
 * Changed
